✨  restructure  de code to make to  easy to understand
