+++ server/package.json
@@ -2,9 +2,10 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node app.js",
+    "dev": "nodemon app.js"
   },
   "keywords": [],
   "author": "",
@@ -16,5 +17,8 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "request": "^2.88.2"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.1"
   }
 }
